taglinkitem

+++ resources/tag.py
@@ -1,7 +1,7 @@
 from flask_smorest import Blueprint, abort
-from schemas import TagSchema
+from schemas import TagSchema, TagAndItemSchema
 from flask.views import MethodView
-from models import TagModel, StoreModel
+from models import TagModel, StoreModel, ItemTags, ItemModel
 
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
@@ -29,6 +29,37 @@ class TagInStore(MethodView):
         return tag
 
 
+@blp.route("/item/<string:item_id>/tag/<string:tag_id>")
+class LinkTagsToItem(MethodView):
+
+    # @blp.arguments()
+    @blp.response(201, TagSchema)
+    def get(self, item_id, tag_id):
+        item = ItemModel.get_or_404(item_id)
+        tag = TagModel.get_or_404(tag_id)
+
+        item.tags.append(tag)
+        try:
+            db.session.add(item)
+            db.session.commit()
+        except SQLAlchemyError as e:
+            abort(500, message=f"error linking tag and item message={e}")
+        return tag
+
+    @blp.response(200, TagSchema)
+    def delete(self, item_id, tag_id):
+        item = ItemModel.get_or_404(item_id)
+        tag = TagModel.get_or_404(tag_id)
+
+        item.tags.remove(tag)
+        try:
+            db.session.add(item)
+            db.session.commit()
+        except SQLAlchemyError as e:
+            abort(500, message=f"error removing tag and item message={e}")
+        return {'message': "item removed from tag ", "item": item, "tag": tag}
+
+
 @blp.route("/tag/<string:tag_id>")
 class Store(MethodView):
 
@@ -37,8 +68,13 @@ class Store(MethodView):
         tag = TagModel.query.get_or_404(tag_id)
         return tag
 
+    @blp.response(202, description="deletes a tag if no item is tagged with it.")
+    @blp.alt_response(404, description="tag not found")
+    @blp.alt_response(400, description="Returns if the tag is assigned to one or more items, the tag is not deleted")
     def delete(self, tag_id):
         tag = StoreModel.query.get_or_404(tag_id)
-        db.session.delete(tag)
-        db.session.commit()
-        return {"message": "tag deleted"}
+        if not tag.items:
+            db.session.delete(tag)
+            db.session.commit()
+            return {"message": "tag deleted"}
+        abort(400, message="Could not delete tag. make sure tag is not associated with any items")
